cmdutil: refactor
